Optionally avoid transmitting sensitive information about a user's system.
This adds the proxyPrivacy option. When set to true, try to avoid
transmitting local host name, port, and time zone. When Polipo is
used with Tor, this information might be useful to an adversary
who wants to identify an anonymous user.

This issue was raised by Gregory Fleischer in his DC 17 presentation,
"Attacking Tor at the Application Layer".

http://pseudo-flaw.net/resource/defcon/attach:/defcon-17-fleischer.pdf
http://pseudo-flaw.net/content/defcon/dc-17-demos/