Don't revalidate objects >= CHUNK_SIZE that aren't in disk cache.
This patch initializes 'last' correctly for non-Range requests (i.e. requests
where 'to' is -1).

This fixes the symptom where Polipo was always revalidating objects with
Content-Length >= CHUNK_SIZE that had not been written to the disk cache;
'objectHasData' was reporting that the object was not in memory nor on
disk (the object wasn't on disk but it *was* in memory).