Cache to disk even when "HTTP://" scheme is uppercase
We have a client that is sending GET requests with a URL containing an
uppercase scheme name. Polipo does cache the instance in memory, but was
not writing it to disk (which prevented it being cached across polipo
restarts).

The client is incorrect because RFC 2616 says in section 3.2.2:

    http_URL = "http:" "//" host [ ":" port ] [ abs_path [ "?" query ]]

But polipo is also incorrect because section 3.2.3 says:

    Comparisons of scheme names MUST be case-insensitive