Fix Polipo ignores cache control values when it has restarted.
The max-age value of Cache-control could be received in a http header.
Polipo web cache proxy should be set this value as the maximum time
that this page is stored in cache. When max-age has elapsed the cache is
stale and it should be refreshed making a new request to the server.

Polipo is saving the cache control values in memory but they are not
saved on disk doing that polipo doesn't know the cache control values when
is restarted and setting them with a default value.

httpPrintCacheControl is the polipo's function which writes the values from a
CacheControlRec object in a buffer. It has two callers:
    - httpWriteRequest which does pass in a CacheControlRec object.
    - httpWriteObjectHeaders which did pass a NULL object rather than
      a CacheControlRec object.

httpWriteObjectHeaders is called when polipo writes the header values on disk
cache so some cache control values was never written on disk. It has been
updated to write cache-control values on disk.

- Updated the call to httpPrintCacheControl from httpWriteObjectHeaders
  passing an updated CacheControlRec object as parameter.
- Updated validateEntry function to fill the memory cache-control values
  when they have been read from disk cache.