Add support for systemd socket activation
Adding systemd support allows spawning polipo in parallel with the services
that depend on polipo.

The caller (spawner) owns the sockets, then spawns polipo and tells polipo
which sockets it has to use. (Before this change, you'd have to spawn polipo,
ask it which sockets it's using, and only then spawn the services that depend
on polipo.)

With socket passing the listening socket is passed to polipo as file
descriptor 3 and the environment variables are set according to:

    LISTEN_FDS=1
    LISTEN_PID=[polipo pid]

before exec()ing polipo